refactored (suggested from code review)
